# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Voitity API",
        "description": "API documentation for Voitity",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/profile/{profile}/messages": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Send a message to a profile",
                "description": "Stores a user message and triggers the AI workflow to generate a reply.",
                "operationId": "9a53cc855d7db2fd9634845cbd660251",
                "parameters": [
                    {
                        "name": "profile",
                        "in": "path",
                        "description": "Profile identifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message"
                                ],
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "Can you summarize my notes?"
                                    },
                                    "chat_id": {
                                        "description": "Existing chat identifier",
                                        "type": "integer",
                                        "example": 12,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Message processed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Message processed successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "chat_id": {
                                                    "type": "integer",
                                                    "example": 42
                                                },
                                                "message_id": {
                                                    "type": "integer",
                                                    "example": 1050
                                                },
                                                "text": {
                                                    "type": "string",
                                                    "example": "Here is the information you requested."
                                                },
                                                "audio_url": {
                                                    "type": "string",
                                                    "example": "https://cdn.example.com/audio/answer.mp3",
                                                    "nullable": true
                                                },
                                                "source": {
                                                    "type": "string",
                                                    "example": "openai"
                                                },
                                                "data": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "202": {
                        "description": "Message stored and processing pending"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "402": {
                        "description": "Subscription limit exceeded"
                    },
                    "404": {
                        "description": "Profile or chat not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "502": {
                        "description": "Answer generation failed"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/profile/{profile}/messages/audio": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Send an audio message to a profile",
                "description": "Transcribes a user audio recording, stores it as a message, and triggers the AI workflow to generate a reply.",
                "operationId": "73a4ae83f788e491ba26facbd8c3f70c",
                "parameters": [
                    {
                        "name": "profile",
                        "in": "path",
                        "description": "Profile identifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "audio"
                                ],
                                "properties": {
                                    "audio": {
                                        "description": "Audio recording to transcribe. Browser WebM/MP4 recordings may be detected as video/webm or video/mp4 containers.",
                                        "type": "file",
                                        "format": "binary"
                                    },
                                    "chat_id": {
                                        "description": "Existing chat identifier",
                                        "type": "integer",
                                        "example": 12,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Audio message processed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Message processed successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "chat_id": {
                                                    "type": "integer",
                                                    "example": 42
                                                },
                                                "message_id": {
                                                    "description": "Answer message identifier",
                                                    "type": "integer",
                                                    "example": 1051
                                                },
                                                "text": {
                                                    "type": "string",
                                                    "example": "Here is the information you requested."
                                                },
                                                "audio_url": {
                                                    "type": "string",
                                                    "example": "https://cdn.example.com/audio/answer.mp3",
                                                    "nullable": true
                                                },
                                                "request_message_id": {
                                                    "description": "Stored user audio message identifier",
                                                    "type": "integer",
                                                    "example": 1050
                                                },
                                                "request_text": {
                                                    "description": "Transcribed text from the uploaded audio",
                                                    "type": "string",
                                                    "example": "Necesito ayuda con mi perfil"
                                                },
                                                "request_audio_url": {
                                                    "type": "string",
                                                    "example": "https://cdn.example.com/messages/audio/1/recording.webm",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "202": {
                        "description": "Audio message stored and processing pending"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "402": {
                        "description": "Subscription limit exceeded"
                    },
                    "404": {
                        "description": "Profile or chat not found"
                    },
                    "422": {
                        "description": "Validation error or empty transcription"
                    },
                    "502": {
                        "description": "Audio transcription failed"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}